feat: begin typescript migration
