+++ test.tsconfig.json
@@ -1,0 +1,9 @@
+{
+    "extends": "./tsconfig.json",
+    "compilerOptions": {
+        "outDir": "testDist",
+        "resolveJsonModule": true,
+        "esModuleInterop": true
+    },
+    "include": ["test/**/*"]
+}
