+++ tsconfig.json
@@ -1,12 +1,20 @@
 {
   "compilerOptions": {
-      "strictNullChecks": true,
-      "target": "es5",
-      "module": "commonjs",
-      "outDir": "dist",
-      "lib": [
-          "es5"
-      ]
+    "target": "es6",
+    "lib": ["es2015", "dom"],
+    "module": "commonjs",
+    "outDir": "dist",
+    "alwaysStrict": true,
+    "incremental": true,
+    "sourceMap": true,
+    "noImplicitAny": true,
+    "removeComments": true,
+    "noLib": false,
+    "preserveConstEnums": true,
+    "experimentalDecorators": true,
+    "declaration": true,
+    "allowJs": true
   },
-  "files": ["index.d.ts", "ts/test.ts"]
+  "files": ["index.ts", "ts/test.ts"],
+  "include": ["lib/**/*", "test/**/*"]
 }
